v 0.5.0
- sql system overhaul
- bug fix

+++ src/main/java/me/kirito5572/commands/main/moderator/MessagePinCommand.java
@@ -12,7 +12,6 @@ import org.jetbrains.annotations.NotNull;
 
 import java.awt.*;
 import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.Objects;
 import java.util.concurrent.TimeUnit;
 
@@ -33,10 +32,21 @@ public class MessagePinCommand implements ICommand {
             }
         }
         String channelId = event.getChannel().getId();
-        try (ResultSet resultSet = mySqlConnector.Select_Query("SELECT * FROM blitz_bot.Pin WHERE channelId=?;", new int[]{mySqlConnector.STRING}, new String[]{channelId})) {
+        MySqlConnector.QueryData queryData = new MySqlConnector.QueryData();
+        queryData.query = "SELECT * FROM blitz_bot.Pin WHERE channelId=?;";
+        queryData.dataType = new int[]{mySqlConnector.STRING};
+        queryData.data = new String[]{channelId};
+        try (ResultSet resultSet = mySqlConnector.Select_Query(queryData)) {
+            if(resultSet == null) {
+                return;
+            }
             if(resultSet.next()) {
                 event.getChannel().deleteMessageById(resultSet.getString("messageId")).queue();
-                mySqlConnector.Insert_Query("DELETE FROM blitz_bot.Pin WHERE channelId=?;" , new int[]{mySqlConnector.STRING}, new String[]{channelId});
+                MySqlConnector.QueryData queryData1 = new MySqlConnector.QueryData();
+                queryData1.query = "DELETE FROM blitz_bot.Pin WHERE channelId=?";
+                queryData1.dataType = new int[]{mySqlConnector.STRING};
+                queryData1.data = new String[]{channelId};
+                mySqlConnector.Insert_Query(queryData1);
                 event.getChannel().sendMessage("핀이 해제되었습니다.").queue(message -> message.delete().queueAfter(10, TimeUnit.SECONDS));
             } else {
                 OptionMapping opt = event.getOption("메세지");
@@ -55,13 +65,11 @@ public class MessagePinCommand implements ICommand {
                 }
                 event.getChannel().sendMessageEmbeds(builder.build()).queue(message -> {
                     String messageId = message.getId();
-                    try {
-                        mySqlConnector.Insert_Query("INSERT INTO blitz_bot.Pin (channelId, messageId) VALUES (?, ?)",
-                                new int[]{mySqlConnector.STRING, mySqlConnector.STRING},
-                                new String[]{channelId, messageId});
-                    } catch (SQLException sqlException) {
-                        sqlException.fillInStackTrace();
-                    }
+                    MySqlConnector.QueryData queryData2 = new MySqlConnector.QueryData();
+                    queryData2.query = "INSERT INTO blitz_bot.Pin (channelId, messageId) VALUES (?, ?)";
+                    queryData2.dataType = new int[]{mySqlConnector.STRING, mySqlConnector.STRING};
+                    queryData2.data = new String[]{channelId, messageId};
+                    mySqlConnector.Insert_Query(queryData2);
                 });
             }
         } catch (Exception e) {
@@ -75,6 +83,7 @@ public class MessagePinCommand implements ICommand {
         return "(관리자 전용)메세지를 고정합니다.";
     }
 
+    @NotNull
     @Override
     public String getInvoke() {
         return "핀";
